requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.11.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
defusedxml==0.7.1
Django==3.2.19
django-allauth==0.58.2
django-ckeditor==6.7.0
django-filter==23.4
django-grappelli==3.0.6
django-js-asset==2.1.0
django-mptt==0.14.0
django-render-partial==0.4
django-solo==2.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
model-bakery==1.17.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.0.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
psycopg2-binary==2.9.3